Allow the user to specify an encoding for the buffer when storing on the server.
The current implementation was using 'utf8' as that is the default when
dealing with Node buffers. However, I wanted to store images taken from
an HTML canvas element via canvas.toDataURL(), which returns a base64
encoded png as a data URL.

Note: I have done a few simple tests, and this code appears to work
with 'utf8' and 'base64' encodings. However, other encodings are
untested and may or may not work (unsure if MongoDB places any
limitations on acceptable char encodings or not).