cleanup stored/uploaded events and further improve chunk tracking
Last remaining change is to try to pass in chunkSum to createWriteStream, so that we don't have to track it in the fileObj. This will involve having the client uploader pass the chunkSum to the access point in query params, and then passing it along to tempstore. Other places that call FS.TempStore.createWriteStream will have to pass it in, too.