Initial commit of changes to support multi-instance systems
The worker package now observes a Mongo backed queue for jobs, which is
handled by a new package job-manager. The collection observers have
been split out into a separate package to enable them to be limited to
a single instance when scaling. Standard packages has also been updated
to include job-manager and collection-observers.